// Frontend/src/api/models/StaffInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ServePOS.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StaffInfo
 */
export interface StaffInfo {
    /**
     * 
     * @type {number}
     * @memberof StaffInfo
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof StaffInfo
     */
    name: string | null;
    /**
     * 
     * @type {Date}
     * @memberof StaffInfo
     */
    lastOperation: Date;
}

/**
 * Check if a given object implements the StaffInfo interface.
 */
export function instanceOfStaffInfo(value: object): value is StaffInfo {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('lastOperation' in value) || value['lastOperation'] === undefined) return false;
    return true;
}

export function StaffInfoFromJSON(json: any): StaffInfo {
    return StaffInfoFromJSONTyped(json, false);
}

export function StaffInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StaffInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'lastOperation': (new Date(json['lastOperation'])),
    };
}

export function StaffInfoToJSON(json: any): StaffInfo {
    return StaffInfoToJSONTyped(json, false);
}

export function StaffInfoToJSONTyped(value?: StaffInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'lastOperation': value['lastOperation'].toISOString(),
    };
}